Data.Attoparsec.Enumerator

Description

Documentation

data ParseError

The context and message from a Fail value.

Constructors

ParseError

Fields

errorContexts :: [String]

errorMessage :: String

class AttoparsecInput a

A class of types which may be consumed by an Attoparsec parser.

Since: 0.3

Minimal complete definition

parseA, feedA, empty, isNull, notEmpty

iterParser :: (AttoparsecInput a, Monad m) => Parser a b -> Iteratee a m b

Convert an Attoparsec Parser into an Iteratee. The parser will be streamed bytes until it returns Done or Fail.

If parsing fails, a ParseError will be thrown with throwError. Use catchError to catch it.